+++ wrapper/R/R_package/hardcoded_wrappers/hardcoded_wrapper.h
@@ -31,11 +31,11 @@ void AT_SPC_get_size_from_filename_R(const char** filename,
 void AT_SPC_read_data_from_filename_R( const char** filename,
 		const int* n,
 		int* depth_step,
-		float* depth_g_cm2,
-		float* E_MeV_u,
-		float* DE_MeV_u,
+		double* depth_g_cm2,
+		double* E_MeV_u,
+		double* DE_MeV_u,
 		int* particle_no,
-		float* fluence_cm2,
+		double* fluence_cm2,
         int* n_bins_read);
 
 #endif /* AT_HARDCODED_WRAPPER_H_ */
